requirements.txt
flask==3.1.0
google-api-python-client==2.166.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.13.3
gunicorn==23.0.0
